Add Filtering of Recorded ResourceTypes
AWS Config pricing relies on two axis: rule-evaluation and configuration-item-recorded.  Allowing filtering of which resourceTypes are to be recorded enables a mechanism to control costs.

This commit adds a new optional variable.  Not setting it defaults to the original/current behaviour.

Setting `recorded_resource_types` will configure an allowlist of which items will be recorded.  Configuration Items found by the rules, but not associated to these resourceTypes will be dropped.